fix output of toString: output the doubles not ints
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@7476 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/qsar/result/DoubleArrayResult.java
@@ -61,7 +61,7 @@ public class DoubleArrayResult implements IDescriptorResult {
         StringBuffer buf = new StringBuffer();
         for (Iterator iterator = array.iterator(); iterator.hasNext();) {
             Double value = (Double) iterator.next();
-            buf.append(value.intValue());
+            buf.append(value.doubleValue());
             if (iterator.hasNext()) buf.append(",");
         }
         return buf.toString();
